requirements.txt
# FastAPI 和 Web 服务
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# AssemblyAI 语音转文本
assemblyai==0.35.0
pyaudio==0.2.14

# LangGraph 和 LangChain
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5

# OpenRouter / LLM 集成
openai==1.54.3  # OpenRouter 使用 OpenAI 兼容 API
httpx==0.27.2
requests==2.32.3

# 环境变量管理
python-dotenv==1.0.1

# 数据处理
pydantic==2.9.2
pydantic-settings==2.6.0

# 异步支持
aiofiles==24.1.0

# 日志和工具
loguru==0.7.2
